feat!: make popup accept any widget as the body (#18)
Popup is now a generic over the body widget, allowing for more complex
widgets to be used as the body of the popup. This change also allows for
the popup to be rendered multiple times without needing to clone it.

Fixes: https://github.com/joshka/tui-popup/issues/16

BREAKING CHANGE: The Popup widget now accepts any widget as the body
instead of just a Text. This allows for more complex widgets to be used
as the body of the popup, such as Paragraph or ScrollView.

Additionally the Popup widget now implements `WidgetRef` and
`StatefulWidgetRef` instead of having a separate PopupWidget type. This
allows for saving a reference to the popup and rendering it multiple
times without needing to clone it. To update your code, replace
`frame.render_widget(popup.to_widget(), area)` with
`frame.render_widget(&popup, area)` and replace
`frame.render_stateful_widget(popup.to_widget(), area, state)`
with `frame.render_stateful_widget_ref(&popup, area, state)`.