feat!: Rename SizedWidgetRef to KnownSize
The trait no longer requires implementors to also implement WidgetRef as
that requirement is moved to the Popup generic bounds.

As a result, this means that a better name was needed. Sized is probably
the right name for this, except there is a well known trait in rust with
the same name and this would be confusing. Instead I went with
`KnownSize`.

BREAKING CHANGE: SizedWidgetRef is now named `KnownSize` and no longer
requires implementing WidgetRef.