chore!: Refactor popup (#49)
These changes were done to relax the dependency on ratatui directly,
instead aiming to depend on ratatui-core.
To do this requires avoiding the unstable widget-ref feature. This is a
preparatory change to support that effort (if it's possible and aligns
to the ratatui modularization work going on in 0.30.0 (currently
pointing at the alpha.0 release).

BREAKING CHANGE: `SizedWidgetRef` is now `KnownSize`. It no longer
requires you to implement `WidgetRef` (this requirement is on the popup
now instead of in the trait). It also no longer requires you to
implement Debug.

`SizedWrapper` is now `KnownSizeWrapper`

The popup no longer implements `WidgetRef` and `StatefulWidgetRef`
directly, instead implements `{Stateful,}Widget for &Popup`

---

- **refactor: Move sized widget code to modules**
  
- **feat: impl SizeWidgetRef for String**
  
- **feat: impl StatefulWidget instead of _Ref**
  - impl StatefulWidget for Popup and &Popup
  - add tests
  - add PartialEq for testing Popups
  - add bacon setup for unit tests
  - remove --color always from bacon.toml
  
- **feat: impl Widget for Popup**
  impl Widget for &Popup and Popup Instead of WidgetRef (ratatui now has
  a blanket impl of WidgetRef for this.
  
  Remove the need for Widgets contained in the popup to implement debug.
  
- **chore: test SizedWrapper, add setters and new()**
  
- **feat!: Rename SizedWidgetRef to KnownSize**
The trait no longer requires implementors to also implement WidgetRef as
  that requirement is moved to the Popup generic bounds.
  
As a result, this means that a better name was needed. Sized is probably
the right name for this, except there is a well known trait in rust with
  the same name and this would be confusing. Instead I went with
  `KnownSize`.
  
  BREAKING CHANGE: SizedWidgetRef is now named `KnownSize` and no longer
  requires implementing WidgetRef.
  

- **feat: Rename SizedWrapper to KnownSizeWrapper.**
  

- **refactor: remove unnecessary generic bounds**
  

- **feat: implement KnownSize trait for Text**
  refactor popup rendering code, examples,