feat(tui-bar-graph)!: support boxed gradients
This patch adds support for boxed gradients in the `BarGraph` widget.
This makes it possible to choose gradients of different types at runtime
without having to change the type of the `BarGraph` struct.

BREAKING CHANGE: The `with_gradient` method now requires a lifetime
on the parameter, and this flows throug to the `BarGraph` struct.
Most usages should continue to work as before.