feat(tui-bar-graph)!: support boxed gradients (#66)
This patch adds support for boxed gradients in the `BarGraph` widget.
This makes it possible to choose gradients of different types at runtime
without having to change the type of the `BarGraph` struct.

BREAKING CHANGE: The `with_gradient` method now requires a lifetime
on the parameter, and this flows throug to the `BarGraph` struct.
Most usages should continue to work as before.

![Braille Rainbow](https://vhs.charm.sh/vhs-1sx9Ht6NzU6e28Cl51jJVv.gif)
![Solid Plasma](https://vhs.charm.sh/vhs-7pWuLtZpzrz1OVD04cMt1a.gif)

<details><summary>More examples</summary>

![Braille Magma](https://vhs.charm.sh/vhs-4RDwcz9DApA90iJYMQXHXd.gif)
![Braille Viridis](https://vhs.charm.sh/vhs-5ylsZAdKGPiHUYboOpZFZL.gif)
![Solid Inferno](https://vhs.charm.sh/vhs-4z1gbmJ50KGz2TPej3mnVf.gif)
![Solid Sinebow](https://vhs.charm.sh/vhs-63aAmMhcfMT8CnWCV20dsn.gif)

</details>