feat(tui-big-text)!: add block octant characters (#117)
The changes in this PR add the following enum variants:

* `PixelSize::QuarterHeight` and `PixelSize::Octant` to `PixelSize`
found in the `tui-big-text` crate.

Documentation and tests are included, and the examples (along with the
VHS tapes and screenshots) have been updated to show the new
**two-row-tall text** styles in action and how they compare to text
rendered with other `PixelSize` settings.

BREAKING CHANGE: New enum variants (see above) added to an enum that is
declared exhaustive.